BS-291 | Vijay | add User Name, Start Date and End Date as part of initialize (#46)
* BS-291 | Vijay | add User Name, Start Date and End Date as part of initialize

* BS-291 | Vijay | resolved bean name
