+++ api/src/main/java/org/openmrs/module/fhirExtension/service/ExportAsyncService.java
@@ -2,14 +2,11 @@ package org.openmrs.module.fhirExtension.service;
 
 import lombok.extern.log4j.Log4j2;
 import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.openmrs.Concept;
 import org.openmrs.api.ConceptService;
 import org.openmrs.api.context.Context;
 import org.openmrs.api.context.UserContext;
 import org.openmrs.module.fhir2.api.dao.FhirTaskDao;
 import org.openmrs.module.fhir2.model.FhirTask;
-import org.openmrs.module.fhir2.model.FhirTaskInput;
-import org.openmrs.module.fhir2.model.FhirTaskOutput;
 import org.openmrs.module.fhirExtension.export.Exporter;
 import org.openmrs.module.fhirExtension.export.anonymise.handler.AnonymiseHandler;
 import org.openmrs.module.fhirExtension.export.anonymise.impl.CorrelationCache;
@@ -24,16 +21,6 @@ import java.util.*;
 @Component
 @Transactional
 public class ExportAsyncService {
-	
-	public static final String DOWNLOAD_URL = "Download URL";
-	
-	public static final String USER_NAME_CONCEPT = "FHIR Export User Name";
-	
-	public static final String START_DATE_CONCEPT = "FHIR Export Start Date";
-	
-	public static final String END_DATE_CONCEPT = "FHIR Export End Date";
-	
-	public static final String ANONYMISE_CONCEPT = "FHIR Export Anonymise Flag";
 	
 	private FhirTaskDao fhirTaskDao;
 	
@@ -56,8 +43,7 @@ public class ExportAsyncService {
 	}
 	
 	@Async("export-fhir-data-threadPoolTaskExecutor")
-	public void export(FhirTask fhirTask, String startDate, String endDate, UserContext userContext, String downloadUrl,
-	        boolean isAnonymise) {
+	public void export(FhirTask fhirTask, String startDate, String endDate, UserContext userContext, boolean isAnonymise) {
 		FhirTask.TaskStatus taskStatus = null;
 		
 		try {
@@ -73,10 +59,6 @@ public class ExportAsyncService {
 			}
 			
 			fileExportService.createZipWithExportedNdjsonFiles(fhirTask.getUuid());
-			FhirTaskOutput fhirTaskOutput = getFhirTaskOutput(fhirTask, downloadUrl);
-			fhirTask.setOutput(Collections.singleton(fhirTaskOutput));
-			Set<FhirTaskInput> fhirTaskInputs = getFhirTaskInputs(fhirTask, startDate, endDate, isAnonymise);
-			fhirTask.setInput(fhirTaskInputs);
 			fileExportService.deleteDirectory(fhirTask.getUuid());
 		}
 		catch (Exception exception) {
@@ -96,36 +78,6 @@ public class ExportAsyncService {
 		fileExportService.createDirectory(fhirTask.getUuid());
 		correlationCache.reset();
 		anonymiseHandler.loadAnonymiserConfig(isAnonymise);
-	}
-	
-	private Set<FhirTaskInput> getFhirTaskInputs(FhirTask fhirTask, String startDate, String endDate, boolean isAnonymise) {
-		FhirTaskInput userNameFhirTaskInput = createFHIRTaskInput(fhirTask, USER_NAME_CONCEPT,	Context.getAuthenticatedUser().getUsername());
-        FhirTaskInput startDateFhirTaskInput = createFHIRTaskInput(fhirTask, START_DATE_CONCEPT, startDate);
-        FhirTaskInput endDateFhirTaskInput = createFHIRTaskInput(fhirTask, END_DATE_CONCEPT, endDate);
-        FhirTaskInput anonymiseFhirTaskInput = createFHIRTaskInput(fhirTask, ANONYMISE_CONCEPT, Boolean.toString(isAnonymise));
-        return new HashSet<>(Arrays.asList(userNameFhirTaskInput, startDateFhirTaskInput, endDateFhirTaskInput, anonymiseFhirTaskInput));
-    }
-	
-	private FhirTaskInput createFHIRTaskInput(FhirTask fhirTask, String conceptName, String valueText) {
-		Concept concept = conceptService.getConceptByName(conceptName);
-		if (concept == null) {
-			throw new RuntimeException("Concept with name " + conceptName + " not found");
-		}
-		FhirTaskInput fhirTaskInput = new FhirTaskInput();
-		fhirTaskInput.setName(conceptName);
-		fhirTaskInput.setTask(fhirTask);
-		fhirTaskInput.setValueText(valueText);
-		fhirTaskInput.setType(concept);
-		return fhirTaskInput;
-	}
-	
-	private FhirTaskOutput getFhirTaskOutput(FhirTask fhirTask, String downloadUrl) {
-		FhirTaskOutput fhirTaskOutput = new FhirTaskOutput();
-		fhirTaskOutput.setName("Download Link Name");
-		fhirTaskOutput.setTask(fhirTask);
-		fhirTaskOutput.setValueText(downloadUrl + "?file=" + fhirTask.getUuid());
-		fhirTaskOutput.setType(conceptService.getConceptByName(DOWNLOAD_URL));
-		return fhirTaskOutput;
 	}
 	
 	private void anonymise(List<IBaseResource> iBaseResources, String resourceType, boolean isAnonymise) {
