+++ api/src/main/java/org/openmrs/module/fhirExtension/service/impl/ExportTaskImpl.java
@@ -2,13 +2,21 @@ package org.openmrs.module.fhirExtension.service.impl;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.time.DateUtils;
+import org.openmrs.Concept;
+import org.openmrs.api.ConceptService;
 import org.openmrs.api.context.Context;
 import org.openmrs.module.fhir2.api.dao.FhirTaskDao;
 import org.openmrs.module.fhir2.model.FhirTask;
+import org.openmrs.module.fhir2.model.FhirTaskInput;
+import org.openmrs.module.fhir2.model.FhirTaskOutput;
 import org.openmrs.module.fhirExtension.service.ExportTask;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
 
 import static org.openmrs.module.fhirExtension.export.Exporter.DATE_FORMAT;
 
@@ -20,12 +28,15 @@ public class ExportTaskImpl implements ExportTask {
 	
 	private FhirTaskDao fhirTaskDao;
 	
-	public ExportTaskImpl(FhirTaskDao fhirTaskDao) {
+	private ConceptService conceptService;
+	
+	public ExportTaskImpl(FhirTaskDao fhirTaskDao, ConceptService conceptService) {
 		this.fhirTaskDao = fhirTaskDao;
+		this.conceptService = conceptService;
 	}
 	
 	@Override
-	public FhirTask getInitialTaskResponse(boolean isAnonymise) {
+	public FhirTask getInitialTaskResponse(String startDate, String endDate, String downloadUrl, boolean isAnonymise) {
 		FhirTask fhirTask = new FhirTask();
 		if (!isAnonymise) {
 			Context.requirePrivilege(NON_ANONYMISE_REQUIRED_PRIVILEGE);
@@ -35,7 +46,10 @@ public class ExportTaskImpl implements ExportTask {
 		fhirTask.setName(logMessage);
 		log.info(logMessage);
 		fhirTask.setIntent(FhirTask.TaskIntent.ORDER);
-		
+		FhirTaskOutput fhirTaskOutput = getFhirTaskOutput(fhirTask, downloadUrl);
+		fhirTask.setOutput(Collections.singleton(fhirTaskOutput));
+		Set<FhirTaskInput> fhirTaskInputs = getFhirTaskInputs(fhirTask, startDate, endDate, isAnonymise);
+		fhirTask.setInput(fhirTaskInputs);
 		fhirTaskDao.createOrUpdate(fhirTask);
 		
 		return fhirTask;
@@ -60,5 +74,35 @@ public class ExportTaskImpl implements ExportTask {
 			    startDateStr);
 		}
 		return validationErrorMessage;
+	}
+	
+	private Set<FhirTaskInput> getFhirTaskInputs(FhirTask fhirTask, String startDate, String endDate, boolean isAnonymise) {
+		FhirTaskInput userNameFhirTaskInput = createFHIRTaskInput(fhirTask, USER_NAME_CONCEPT,	Context.getAuthenticatedUser().getUsername());
+		FhirTaskInput startDateFhirTaskInput = createFHIRTaskInput(fhirTask, START_DATE_CONCEPT, startDate);
+		FhirTaskInput endDateFhirTaskInput = createFHIRTaskInput(fhirTask, END_DATE_CONCEPT, endDate);
+		FhirTaskInput anonymiseFhirTaskInput = createFHIRTaskInput(fhirTask, ANONYMISE_CONCEPT, Boolean.toString(isAnonymise));
+		return new HashSet<>(Arrays.asList(userNameFhirTaskInput, startDateFhirTaskInput, endDateFhirTaskInput, anonymiseFhirTaskInput));
+	}
+	
+	private FhirTaskInput createFHIRTaskInput(FhirTask fhirTask, String conceptName, String valueText) {
+		Concept concept = conceptService.getConceptByName(conceptName);
+		if (concept == null) {
+			throw new RuntimeException("Concept with name " + conceptName + " not found");
+		}
+		FhirTaskInput fhirTaskInput = new FhirTaskInput();
+		fhirTaskInput.setName(conceptName);
+		fhirTaskInput.setTask(fhirTask);
+		fhirTaskInput.setValueText(valueText);
+		fhirTaskInput.setType(concept);
+		return fhirTaskInput;
+	}
+	
+	private FhirTaskOutput getFhirTaskOutput(FhirTask fhirTask, String downloadUrl) {
+		FhirTaskOutput fhirTaskOutput = new FhirTaskOutput();
+		fhirTaskOutput.setName("Download Link Name");
+		fhirTaskOutput.setTask(fhirTask);
+		fhirTaskOutput.setValueText(downloadUrl + "?file=" + fhirTask.getUuid());
+		fhirTaskOutput.setType(conceptService.getConceptByName(DOWNLOAD_URL));
+		return fhirTaskOutput;
 	}
 }
