+++ api/src/test/java/org/openmrs/module/fhirExtension/service/impl/ExportTaskTest.java
@@ -5,8 +5,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.openmrs.Concept;
 import org.openmrs.Person;
 import org.openmrs.User;
+import org.openmrs.api.ConceptService;
 import org.openmrs.api.context.Context;
 import org.openmrs.module.fhir2.api.dao.FhirTaskDao;
 import org.openmrs.module.fhir2.model.FhirTask;
@@ -19,12 +21,17 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(Context.class)
 @PowerMockIgnore("javax.management.*")
 public class ExportTaskTest {
+	
+	@Mock
+	private ConceptService conceptService;
 	
 	@Mock
 	private FhirTaskDao fhirTaskDao;
@@ -43,10 +50,17 @@ public class ExportTaskTest {
 	@Test
 	public void shouldCreateFhirTask_whenRequestedForPatientDataExport() {
 		when(fhirTaskDao.createOrUpdate(any(FhirTask.class))).thenReturn(mockFhirTask());
-		
-		FhirTask initialTaskResponse = exportTask.getInitialTaskResponse(false);
+		when(conceptService.getConceptByName(any())).thenReturn(new Concept());
+		FhirTask initialTaskResponse = exportTask.getInitialTaskResponse("2023-05-01", "2023-05-31", "http://dummyUrl",
+		    false);
 		
 		assertNotNull(initialTaskResponse);
+		verify(conceptService, times(1)).getConceptByName("Download URL");
+		verify(conceptService, times(1)).getConceptByName("FHIR Export User Name");
+		verify(conceptService, times(1)).getConceptByName("FHIR Export Start Date");
+		verify(conceptService, times(1)).getConceptByName("FHIR Export End Date");
+		verify(conceptService, times(1)).getConceptByName("FHIR Export Anonymise Flag");
+		verify(fhirTaskDao, times(1)).createOrUpdate(any(FhirTask.class));
 		assertEquals(FhirTask.TaskStatus.ACCEPTED, initialTaskResponse.getStatus());
 	}
