+++ omod/src/main/java/org/openmrs/module/fhirExtension/web/ExportController.java
@@ -46,9 +46,8 @@ public class ExportController extends BaseRestController {
 			response.add("error", validationErrorMessage);
 			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 		}
-		FhirTask fhirTask = exportTask.getInitialTaskResponse(isAnonymise);
-		exportAsyncService.export(fhirTask, startDate, endDate, Context.getUserContext(),
-		    ServletUriComponentsBuilder.fromCurrentContextPath().toUriString() + FILE_DOWNLOAD_URI, isAnonymise);
+		FhirTask fhirTask = exportTask.getInitialTaskResponse(startDate, endDate, ServletUriComponentsBuilder.fromCurrentContextPath().toUriString() + FILE_DOWNLOAD_URI, isAnonymise);
+		exportAsyncService.export(fhirTask, startDate, endDate, Context.getUserContext(), isAnonymise);
 		return new ResponseEntity<>(getFhirTaskUri(fhirTask), HttpStatus.ACCEPTED);
 	}
