+++ omod/src/test/java/org/openmrs/module/fhirExtension/web/ExportControllerTest.java
@@ -60,8 +60,8 @@ public class ExportControllerTest {
 	
 	@Test
 	public void shouldGetFhirTaskUrl_whenFhirExportCalled() {
-		doNothing().when(exportAsyncService).export(any(), any(), any(), any(), any(), anyBoolean());
-		when(exportTask.getInitialTaskResponse(true)).thenReturn(mockFhirTask());
+		doNothing().when(exportAsyncService).export(any(), any(), any(), any(), anyBoolean());
+		when(exportTask.getInitialTaskResponse(any(), any(), any(), anyBoolean())).thenReturn(mockFhirTask());
 		when(exportTask.validateParams("2023-05-01", "2023-05-31")).thenReturn(null);
 		ResponseEntity<SimpleObject> responseEntity = exportController.export("2023-05-01", "2023-05-31", true);
 		SimpleObject simpleObject = responseEntity.getBody();
@@ -73,8 +73,8 @@ public class ExportControllerTest {
 	
 	@Test
 	public void shouldGetBadRequest_whenFhirExportCalledWithInvalidDateFormat() {
-		doNothing().when(exportAsyncService).export(any(), any(), any(), any(), any(), anyBoolean());
-		when(exportTask.getInitialTaskResponse(true)).thenReturn(mockFhirTask());
+		doNothing().when(exportAsyncService).export(any(), any(), any(), any(), anyBoolean());
+		when(exportTask.getInitialTaskResponse(any(), any(), any(), anyBoolean())).thenReturn(mockFhirTask());
 		when(exportTask.validateParams("2023-05-AB", "2023-05-31")).thenReturn("Invalid Date Format [yyyy-mm-dd]");
 		ResponseEntity<SimpleObject> responseEntity = exportController.export("2023-05-AB", "2023-05-31", true);
 		SimpleObject simpleObject = responseEntity.getBody();
@@ -84,7 +84,7 @@ public class ExportControllerTest {
 	
 	@Test
 	public void shouldThrowException_whenLoggedInUserDoesNotHavePrivilegeToExportNonAnonymisedData() {
-		when(exportTask.getInitialTaskResponse(false)).thenThrow( new ContextAuthenticationException( "Privileges required: Export Non Anonymised Patient Data"));
+		when(exportTask.getInitialTaskResponse(any(), any(), any(),anyBoolean())).thenThrow( new ContextAuthenticationException( "Privileges required: Export Non Anonymised Patient Data"));
 		when(exportTask.validateParams("2023-05-AB", "2023-05-31")).thenReturn(null);
 		assertThrows(ContextAuthenticationException.class, () -> exportController.export("2023-05-AB", "2023-05-31", false));
 	}
